// src/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
      api_keys: {
        Row: {
          created_at: string | null
          key: string
          organization_id: string
          updated_at: string | null
        }
        Insert: {
          created_at?: string | null
          key: string
          organization_id: string
          updated_at?: string | null
        }
        Update: {
          created_at?: string | null
          key?: string
          organization_id?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "api_keys_organization_id_fkey"
            columns: ["organization_id"]
            isOneToOne: true
            referencedRelation: "organizations"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      sitemap_logs: {
        Row: {
          created_at: string | null
          does_sitemap_contain_errors: boolean
          fetching_duration: number
          id: string
          mode: string | null
          number_of_sitemap_fetched: number
          organization_id: string
          source: string | null
          total_pages_in_sitemaps: number
          url: string
        }
        Insert: {
          created_at?: string | null
          does_sitemap_contain_errors: boolean
          fetching_duration: number
          id?: string
          mode?: string | null
          number_of_sitemap_fetched: number
          organization_id: string
          source?: string | null
          total_pages_in_sitemaps: number
          url: string
        }
        Update: {
          created_at?: string | null
          does_sitemap_contain_errors?: boolean
          fetching_duration?: number
          id?: string
          mode?: string | null
          number_of_sitemap_fetched?: number
          organization_id?: string
          source?: string | null
          total_pages_in_sitemaps?: number
          url?: string
        }
        Relationships: [
          {
            foreignKeyName: "sitemap_logs_organization_id_fkey"
            columns: ["organization_id"]
            isOneToOne: false
            referencedRelation: "organizations"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}